move replication-actions component into core addon so that secondary recovery works
