+++ ui/lib/core/addon/components/replication-action-promote.js
@@ -1,0 +1,6 @@
+import Actions from './replication-actions-single';
+import layout from '../templates/components/replication-action-promote';
+
+export default Actions.extend({
+  layout,
+});
