+++ ui/lib/core/addon/components/replication-actions.js
@@ -1,7 +1,8 @@
 import { alias } from '@ember/object/computed';
 import Component from '@ember/component';
 import { computed } from '@ember/object';
-import ReplicationActions from 'replication/mixins/replication-actions';
+import ReplicationActions from 'core/mixins/replication-actions';
+import layout from '../templates/components/replication-actions';
 
 const DEFAULTS = {
   token: null,
@@ -14,13 +15,10 @@ const DEFAULTS = {
 };
 
 export default Component.extend(ReplicationActions, DEFAULTS, {
+  layout,
   replicationMode: null,
   model: null,
   cluster: alias('model'),
-
-  didReceiveAttrs() {
-    this._super(...arguments);
-  },
 
   reset() {
     if (!this || this.isDestroyed || this.isDestroying) {
