+++ ui/lib/core/addon/helpers/replication-action-for-mode.js
