+++ ui/lib/core/addon/templates/components/replication-action-promote.hbs
