+++ ui/lib/core/addon/templates/components/replication-action-update-primary.hbs
