+++ ui/lib/core/app/components/replication-action-promote.js
@@ -1,0 +1,1 @@
+export { default } from 'core/components/replication-action-promote';
