+++ ui/lib/core/app/components/replication-action-update-primary.js
@@ -1,0 +1,1 @@
+export { default } from 'core/components/replication-action-update-primary';
