+++ ui/lib/core/app/components/replication-actions-single.js
@@ -1,0 +1,1 @@
+export { default } from 'core/components/replication-actions-single';
