+++ ui/lib/core/app/helpers/replication-action-for-mode.js
@@ -1,0 +1,1 @@
+export { default } from 'core/helpers/replication-action-for-mode';
