+++ ui/lib/core/app/mixins/replication-actions.js
@@ -1,0 +1,1 @@
+export { default } from 'core/mixins/replication-actions';
