+++ ui/lib/replication/addon/components/replication-action-disable.js
@@ -1,3 +1,3 @@
-import Actions from './replication-actions-single';
+import Actions from 'core/components/replication-actions-single';
 
 export default Actions.extend();
