--- ui/lib/replication/addon/components/replication-action-promote.js
@@ -1,3 +1,0 @@
-import Actions from './replication-actions-single';
-
-export default Actions.extend();
