+++ ui/tests/integration/components/replication-actions-test.js
@@ -4,13 +4,8 @@ import Service from '@ember/service';
 import { module, test } from 'qunit';
 import { setupRenderingTest } from 'ember-qunit';
 import { click, fillIn, blur, render, find } from '@ember/test-helpers';
-import { camelize } from '@ember/string';
-import engineResolverFor from 'ember-engines/test-support/engine-resolver-for';
-
 import hbs from 'htmlbars-inline-precompile';
 import sinon from 'sinon';
-
-const resolver = engineResolverFor('replication');
 
 const storeStub = Service.extend({
   callArgs: null,
@@ -32,7 +27,7 @@ const routerService = Service.extend({
 });
 
 module('Integration | Component | replication actions', function(hooks) {
-  setupRenderingTest(hooks, { resolver });
+  setupRenderingTest(hooks);
 
   hooks.beforeEach(function() {
     run(() => {
@@ -99,7 +94,6 @@ module('Integration | Component | replication actions', function(hooks) {
           mode: clusterMode,
           modeForUrl: clusterMode,
         },
-        [`can${camelize(action)}`]: true,
         reload() {
           return resolve();
         },
